(Assessment 2) - Database - Insert and Read(View) data

+++ app/src/main/java/com/mopro/rumusdasarmatermatika/model/BangunPersegi.kt
@@ -1,7 +1,5 @@
 package com.mopro.rumusdasarmatermatika.model
 
-import java.math.BigInteger
-
 data class BangunPersegi(
-    val sisi: BigInteger
+    val sisi: Long
 )
