+++ app/src/main/java/com/mopro/rumusdasarmatermatika/model/BangunPersegiPanjang.kt
@@ -1,8 +1,7 @@
 package com.mopro.rumusdasarmatermatika.model
 
-import java.math.BigInteger
 
 data class BangunPersegiPanjang(
-    val panjang: BigInteger,
-    val lebar: BigInteger
+    val panjang: Long,
+    val lebar: Long
 )
