+++ app/src/main/java/com/mopro/rumusdasarmatermatika/model/BangunSegitiga.kt
@@ -1,6 +1,6 @@
 package com.mopro.rumusdasarmatermatika.model
 
 data class BangunSegitiga(
-    val alas: Float,
-    val tinggi: Float
+    val alas: Long,
+    val tinggi: Long
 )
