+++ app/src/main/java/com/mopro/rumusdasarmatermatika/model/LuasPersegi.kt
@@ -3,6 +3,6 @@ package com.mopro.rumusdasarmatermatika.model
 import java.math.BigInteger
 
 data class LuasPersegi(
-    val hasil: BigInteger
+    val hasil: Long
 
 )
