+++ app/src/main/java/com/mopro/rumusdasarmatermatika/model/LuasPersegiPanjang.kt
@@ -1,7 +1,5 @@
 package com.mopro.rumusdasarmatermatika.model
 
-import java.math.BigInteger
-
 data class LuasPersegiPanjang(
-    val hasil: BigInteger
+    val hasil: Long
 )
